Flagged the template as specific to Windows 8 devices
Important: that change is NOT compatible with the Factory v1!
Updating the template to this version in the Factory would cause
the template to disappear from the list of templates that are
available.

In version 2, this update means that:
1. the template won't be displayed in the list of templates
available for regular tablets/desktops. That's good because the
template needs a Windows 8 machine.

2. the template will be displayed in the list of templates available
for Windows 8 tablets/desktops. The Factory now only renders the
templates that explicitly support these devices when this device is
selected.